cloud/requirements.txt
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pdf2image==1.3.1
Pillow==5.4.0
python-dotenv==0.10.1
reportlab==3.5.12
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
